Update README.md
The code allows the user to log in, comparing the password inputed by the user to the usernames and passwords on file.

After logging in, the user can crreate new bank account, check account details or log out.

In the process of creating a new bank account for the user, the program automatically generates a unique 10-digit account number for the user.

The program fetches the account details of an already registered user if the user is a registered user who requests for his or her account balance.

When the user logs out, their session file is also deleted.